Specify non-ISO8601 constraining behaviour in addition and conversion
This adds overrides of CalendarDateAdd and CalendarDateToISO that
specify the general algorithm for constraining dates in calendars other
than ISO8601. They are prose descriptions; the actual steps for each
calendar remain implementation-defined.

Some of the steps do not actually apply to any calendar that CLDR
currently describes, but are still needed for a full generalization.

Closes: #32