Add table with calendar descriptions
Temporal moves the "Calendar types" section from ECMA-402 to ECMA-262.
This adds a new "Calendar types" section to ECMA-402, that refers to the
ECMA-262 one, and contains additional information for implementations
that conform to ECMA-402.

In particular, this adds a table with high-level descriptions of each
supported calendar system, 1 to 2 sentences long, so that
implementations can align on which calendar they refer to exactly.

"islamic" and "islamic-rgsa" are included, pending #32.

Closes: #2