Calendars islamic and islamic-rgsa should fall back to other values
These two calendars have been present on the web for a long time, but
their behaviour does not match any Hijri calendar algorithm in use in
the real world outside of software. So anyone using these calendar IDs
is probably not getting what they expect to get.

When creating a DateTimeFormat with one of these calendar IDs, we
specify a fallback, which is implementation- and locale-defined, but
must be one of the available non-alias calendar IDs. This way,
implementations could choose to fall back to one of the other Hijri
calendars depending on the locale. resolvedOptions().calendar will show
which fallback was chosen.

They are removed from AvailableCalendars, which removes them from
supportedValuesOf("calendar") and from Temporal APIs. Note that the
result of resolvedOptions().calendar can still be passed to any Temporal
API.

See: #29