Editorial: Add definitions for "arithmetic year" and "epoch year"
The definitions link to the description of the [[Year]] field in
Calendar Date Records, where arithmetic year and epoch year are
explained.

Also adjusts the text in a few places to avoid repetition now that we
have linked definitions, and adds a cross-reference to the table of
epoch years.

See: #51