Editorial: Clarify that eras in Table 2 are the currently known ones
Refer to UTS 35 as the canonical source for this data. Add Part 4: Dates
to the list of normative references, in order to clarify that
implementations do not have to omit newly added eras in order to be
conformant, as well as for the note below about updating on a best-
effort basis.