Editorial: Address high-priority points from TG2 editor review (#97)
* Editorial: Add definitions for "arithmetic year" and "epoch year"

The definitions link to the description of the [[Year]] field in
Calendar Date Records, where arithmetic year and epoch year are
explained.

Also adjusts the text in a few places to avoid repetition now that we
have linked definitions, and adds a cross-reference to the table of
epoch years.

See: #51

* Editorial: Clarify that calendars not in Table 1 are implementation-defined

Implementations should not be constrained to just the calendar types in
Table 1, for future extensibility. In any abstract operations that seem
to assume that all calendar types are listed in Table 1, add language
clarifying what happens if they are not.

See: #51
Closes: #94

* Editorial: Clarify that eras in Table 2 are the currently known ones

Refer to UTS 35 as the canonical source for this data. Add Part 4: Dates
to the list of normative references, in order to clarify that
implementations do not have to omit newly added eras in order to be
conformant, as well as for the note below about updating on a best-
effort basis.

* Editorial: Update table-eras caption

Suggestion from code review.

Co-Authored-By: Richard Gibson <richard.gibson@gmail.com>

---------

Co-authored-by: Richard Gibson <richard.gibson@gmail.com>