Bug fix: remove spurious normative change introduced in conversion from prose algorithm descriptions
Bug fix: The replacement of the prose algorithm for `NonISODateUntil`
carried out by 0a7577f386d6f919a371a9037727c53e7c8b8597 introduced an
unwanted normative change: whereas the prose algorithm lexicographically compared the
month codes in the candidate date and the target date when determining
whether the candidate date "surpasses" the target date without
constraining the month given in the candidate date to one that exists in
the year in the candidate date, the new algorithm steps constrains the
month code of the candidate date, converts it to an ordinal month, and
then compares that ordinal month to the month of the target date.

The accidental normative change is only visible when
1. The requested calendar is a lunisolar calendar that uses the
   ~skip-forward~ leap to common month transformation (in practice, only
   the hebrew calendar)
2. The date given by the `_two_` parameter of `NonISODateUntil` is
   before the date given in the `_one_` parameter

This commit rewrites NonISODateSurpasses to compare the month code
lexicographically.

Co-Authored-By: Philip Chimento <pchimento@igalia.com>